referee/requirements.txt
flask
python-chess
flask-cors
